+++ src/main/java/com/OEzoa/OEasy/api/member/MemberController.java
@@ -1,23 +1,18 @@
 package com.OEzoa.OEasy.api.member;
 
 import com.OEzoa.OEasy.application.member.MemberService;
+import com.OEzoa.OEasy.application.member.TokenValidator;
 import com.OEzoa.OEasy.application.member.dto.MemberDTO;
 import com.OEzoa.OEasy.application.member.dto.MemberSignUpDTO;
 import com.OEzoa.OEasy.application.member.mapper.MemberMapper;
 import com.OEzoa.OEasy.domain.member.Member;
 import com.OEzoa.OEasy.domain.member.MemberRepository;
-import com.OEzoa.OEasy.domain.member.MemberToken;
 import com.OEzoa.OEasy.domain.member.MemberTokenRepository;
-import com.OEzoa.OEasy.util.JwtTokenProvider;
 import com.OEzoa.OEasy.util.s3Bucket.FileUploader;
-import io.jsonwebtoken.ExpiredJwtException;
-import io.jsonwebtoken.MalformedJwtException;
-import io.jsonwebtoken.UnsupportedJwtException;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CookieValue;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -38,9 +33,7 @@ public class MemberController {
     private final MemberRepository memberRepository;
     private final FileUploader fileUploader;
     private final MemberTokenRepository memberTokenRepository;
-
-    @Autowired
-    private JwtTokenProvider jwtTokenProvider;
+    private final TokenValidator tokenValidator;
 
     // 일반 회원가입
     @PostMapping("/signup")
@@ -76,7 +69,8 @@ public class MemberController {
     )
     public ResponseEntity<?> getProfile(@CookieValue(name = "accessToken", required = false) String accessToken) {
         try {
-            MemberDTO memberDTO = validateMemberAccessToken(accessToken);
+            Member member = tokenValidator.validateAccessTokenAndReturnMember(accessToken);
+            MemberDTO memberDTO = MemberMapper.toDto(member);
             return ResponseEntity.ok(memberDTO);
         } catch (Exception e) {
             return ResponseEntity.status(401).body(e.getMessage());
@@ -96,7 +90,7 @@ public class MemberController {
     public ResponseEntity<?> updateNickname(@RequestParam String newNickname,
                                             @CookieValue(name = "accessToken", required = false) String accessToken) {
         try {
-            Member member = validateMemberAccessTokenAndReturnMember(accessToken);
+            Member member = tokenValidator.validateAccessTokenAndReturnMember(accessToken);
             member = member.toBuilder().nickname(newNickname).build();
             memberRepository.save(member);
             return ResponseEntity.ok("닉네임 변경 성공");
@@ -118,7 +112,7 @@ public class MemberController {
     public ResponseEntity<?> updateProfilePicture(@RequestParam String imageName, @RequestParam String imageUri,
                                                   @CookieValue(name = "accessToken", required = false) String accessToken) {
         try {
-            Member member = validateMemberAccessTokenAndReturnMember(accessToken);
+            Member member = tokenValidator.validateAccessTokenAndReturnMember(accessToken);
             // S3 버킷에 이미지 업로드
             String s3ImageUrl = fileUploader.uploadImage(imageName, imageUri);
             member = member.toBuilder().memberImage(s3ImageUrl).build();
@@ -127,53 +121,6 @@ public class MemberController {
         } catch (Exception e) {
             return ResponseEntity.badRequest().body("프로필 사진 변경 실패: " + e.getMessage());
         }
-    }
-
-    // 액세스 토큰 검증 및 사용자 정보 조회 메서드
-    private MemberDTO validateMemberAccessToken(String accessToken) throws Exception {
-        return MemberMapper.toDto(validateMemberAccessTokenAndReturnMember(accessToken));
-    }
-
-    // 액세스 토큰 검증 및 Member 객체 조회 메서드
-    private Member validateMemberAccessTokenAndReturnMember(String accessToken) throws Exception {
-        if (accessToken == null) {
-            throw new IllegalArgumentException("인증되지 않은 사용자입니다. 액세스 토큰이 존재하지 않습니다.");
-        }
-
-        try {
-            // JWT 토큰에서 사용자 ID 추출
-            Long memberId = jwtTokenProvider.getMemberIdFromToken(accessToken);
-            return memberRepository.findById(memberId)
-                    .orElseThrow(() -> new IllegalArgumentException("회원 정보를 찾을 수 없습니다. 잘못된 사용자 ID입니다."));
-        } catch (ExpiredJwtException e) {
-            // 토큰 만료 시 새로운 토큰 발급 로직 추가
-            String refreshToken = getRefreshTokenForMember(accessToken);
-            if (refreshToken == null) {
-                throw new IllegalStateException("토큰이 만료되었습니다. 다시 로그인 해 주세요.", e);
-            }
-            String newAccessToken = jwtTokenProvider.generateToken(jwtTokenProvider.getMemberIdFromToken(refreshToken));
-            return memberRepository.findById(jwtTokenProvider.getMemberIdFromToken(newAccessToken))
-                    .orElseThrow(() -> new IllegalArgumentException("회원 정보를 찾을 수 없습니다. 잘못된 사용자 ID입니다."));
-        } catch (UnsupportedJwtException e) {
-            throw new IllegalArgumentException("지원하지 않는 JWT 토큰입니다.", e);
-        } catch (MalformedJwtException e) {
-            throw new IllegalArgumentException("잘못된 JWT 토큰 형식입니다.", e);
-        } catch (Exception e) {
-            throw new IllegalStateException("유효하지 않은 토큰입니다.", e);
-        }
-    }
-
-    // 토큰 갱신 메서드 추가
-    private String getRefreshTokenForMember(String accessToken) {
-        // 토큰에서 사용자 ID 추출
-        Long memberId = jwtTokenProvider.getMemberIdFromToken(accessToken);
-        MemberToken memberToken = memberTokenRepository.findByMemberId(memberId)
-                .orElseThrow(() -> new IllegalArgumentException("회원 정보를 찾을 수 없습니다. 잘못된 사용자 ID입니다."));
-
-        if (jwtTokenProvider.isTokenExpired(memberToken.getRefreshToken())) {
-            throw new IllegalStateException("리프레시 토큰이 만료되었습니다. 다시 로그인 해 주세요.");
-        }
-        return memberToken.getRefreshToken();
     }
 
 }
